Split types into seperate files
