+++ src/GoatTrip.RestApi/Services/ILocationRetrievalService.cs
@@ -1,0 +1,7 @@
+﻿namespace GoatTrip.RestApi.Services {
+    using Models;
+
+    public interface ILocationRetrievalService {
+        LocationModel Get(string id);
+    }
+}
