+++ src/GoatTrip.RestApi/Services/ILocationSearchPostcodeService.cs
@@ -1,0 +1,8 @@
+﻿namespace GoatTrip.RestApi.Services {
+    using System.Collections.Generic;
+    using Models;
+
+    public interface ILocationSearchPostcodeService {
+        IEnumerable<LocationGroupModel> SearchByPostcode(string postcodeQuery);
+    }
+}
