+++ src/GoatTrip.RestApi/Services/LocationSearchService.cs
@@ -1,73 +1,9 @@
-﻿
 namespace GoatTrip.RestApi.Services {
-
-    using DAL.DTOs;
     using System.Collections.Generic;
     using System.Linq;
     using DAL;
+    using DAL.DTOs;
     using Models;
-
-    public class LocationRetrievalService
-        : ILocationRetrievalService {
-        private ILocationIdEncoder _encoder;
-        private ILocationRepository _repository;
-
-        public LocationRetrievalService(ILocationRepository repository, ILocationIdEncoder encoder) {
-            _repository = repository;
-            _encoder = encoder;
-        }
-
-        public LocationModel Get(string id) {
-            var decodedId = _encoder.Decode(id);
-            var location = _repository.Get(decodedId);
-            return new LocationModel(location);
-        }
-
-    }
-
-    public abstract class LocationSearchBaseService {
-
-        protected LocationSearchBaseService(ILocationQueryValidator queryValidator) {
-            _queryValidator = queryValidator;
-        }
-
-        protected void ValidateAndThrow(string query) {
-            if (!_queryValidator.IsValid(query))
-                throw new InvalidLocationQueryException();
-        }
-
-        private readonly ILocationQueryValidator _queryValidator;
-    }
-
-    public class LocationSearchPostcodeService
-        : LocationSearchBaseService, ILocationSearchPostcodeService {
-        private readonly ILocationRepository _repository;
-        private readonly ILocationQuerySanitiser _postCodeSanitiser;
-
-        public LocationSearchPostcodeService(ILocationRepository repository,
-            ILocationQueryValidator queryValidator, ILocationQuerySanitiser postCodeSanitiser)
-            : base(queryValidator) {
-            _repository = repository;
-            _postCodeSanitiser = postCodeSanitiser;
-        }
-
-        public IEnumerable<LocationGroupModel> SearchByPostcode(string postcodeQuery) {
-            ValidateAndThrow(postcodeQuery);
-            var sanitisedQuery = _postCodeSanitiser.Sanitise(postcodeQuery);
-
-            var results = _repository.FindLocations(sanitisedQuery);
-
-            var groupedResult = Group(results);
-
-            return groupedResult;
-        }
-
-        private IEnumerable<LocationGroupModel> Group(IEnumerable<Location> results) {
-            var locations = results.Select(l => new LocationModel(l));
-            return locations.GroupBy(l => l.Postcode)
-                .Select(g => new LocationGroupModel(g.First().GroupDescription, g.Count(), "/locations/search/" + g.First().GroupDescription));
-        }
-    }
 
     public class LocationSearchService
         : LocationSearchBaseService, ILocationSearchService {
@@ -122,11 +58,5 @@ namespace GoatTrip.RestApi.Services {
         private readonly ILocationIdEncoder _encoder;
         private readonly ILocationQueryFields _locationQueryFields;
         private readonly ILocationGroupRepository _groupRepository;
-    }
-
-    public static class IEnumerableOfLocationGroupExtensions {
-        public static bool HasSingleGroup(this IEnumerable<LocationGroup> operand) {
-            return operand.Count() == 1;
-        }
     }
 }
